Rename files to use UpperCamelCaseCluster/Logic for group-key-mgmt-server
