+++ src/app/clusters/group-key-mgmt-server/GroupKeyManagementCluster.cpp
@@ -16,7 +16,7 @@
  */
 
 #include <app-common/zap-generated/cluster-objects.h>
-#include <app/clusters/group-key-mgmt-server/group-key-mgmt-cluster.h>
+#include <app/clusters/group-key-mgmt-server/GroupKeyManagementCluster.h>
 #include <app/server-cluster/AttributeListBuilder.h>
 #include <app/server-cluster/DefaultServerCluster.h>
 #include <app/server/Server.h>
